--- src/main/java/site/dittotrip/ditto_trip/review/comment/service/CommentSaveService.java
@@ -1,37 +1,0 @@
-package site.dittotrip.ditto_trip.review.comment.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import site.dittotrip.ditto_trip.review.comment.domain.Comment;
-import site.dittotrip.ditto_trip.review.comment.domain.dto.save.CommentSaveReq;
-import site.dittotrip.ditto_trip.review.comment.repository.CommentRepository;
-import site.dittotrip.ditto_trip.review.domain.Review;
-import site.dittotrip.ditto_trip.review.repository.ReviewRepository;
-import site.dittotrip.ditto_trip.user.domain.User;
-
-import java.util.NoSuchElementException;
-import java.util.Optional;
-
-@Service
-@Transactional(readOnly = false)
-@RequiredArgsConstructor
-public class CommentSaveService {
-
-    private final ReviewRepository reviewRepository;
-    private final CommentRepository commentRepository;
-
-    public void saveComment(Long reviewId, Long parentCommentId, User user,
-                            CommentSaveReq commentSaveReq) {
-        Review review = reviewRepository.findById(reviewId).orElseThrow(NoSuchElementException::new);
-
-        Comment parentComment = null;
-        if (parentCommentId != null) {
-            parentComment = commentRepository.findById(parentCommentId).orElseThrow(NoSuchElementException::new);
-        }
-
-        Comment comment = new Comment(commentSaveReq.getBody(), user, review, parentComment);
-        commentRepository.save(comment);
-    }
-
-}
